guardado al funcionar debug
