+++ miniproyecto/lib/main.dart
@@ -1,80 +1,30 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
+import './src/navigation.dart';
 import 'package:provider/provider.dart';
+import './src/crud.dart';
+import './src/dashboard.dart';
+import './src/dataInput.dart';
+import './src/settings.dart';
+
 
 void main() async {
-  WidgetsFlutterBinding.ensureInitialized();
+  WidgetsFlutterBinding.ensureInitialized(); //para asegurar q flutter esté correctamente inicializado
 
-  await dotenv.load(fileName: ".env");
+  await dotenv.load(
+    fileName: ".env",
+  ); //caga var entorno desde .env (para info sensible)
 
-  String supabaseUrl = dotenv.env['SUPABASE_URL'] ?? '';
-  String supabaseAnonKey = dotenv.env['SUPABASE_ANON_KEY'] ?? '';
+  String supabaseUrl = dotenv.env['SUPABASE_URL'] ?? ''; //URL
+  String supabaseAnonKey =
+      dotenv.env['SUPABASE_ANON_KEY'] ??
+      ''; //obtiene clave anónima de supabase desde el archivo .env
 
-  await Supabase.initialize(url: supabaseUrl, anonKey: supabaseAnonKey);
+  await Supabase.initialize(
+    url: supabaseUrl,
+    anonKey: supabaseAnonKey,
+  ); //inicializa conexión con supabase
 
   runApp(MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  // This widget is the root of your application.
-  @override
-  Widget build(BuildContext context) {
-    return ChangeNotifierProvider(
-      create: (context) => MyAppState(),
-      child: MaterialApp(
-        debugShowCheckedModeBanner: false,
-        title: 'PharmaStock',
-        theme: ThemeData(
-          useMaterial3: true,
-          colorScheme: ColorScheme.fromSeed(
-            seedColor: const Color.fromARGB(255, 2, 140, 99),
-          ),
-        ),
-        home: MyHomePage(),
-      ),
-    );
-  }
-}
-
-class MyAppState extends ChangeNotifier {
-  //funciones de la lógica
-  //notifyListeners();
-}
-
-class MyHomePage extends StatefulWidget {
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  var selectedIndex = 0;
-
-  @override
-  Widget build(BuildContext context) {
-    var colorScheme = Theme.of(context).colorScheme;
-
-    Widget page;
-    switch (selectedIndex) {
-      case 0:
-        page = Placeholder();
-        break;
-      case 1:
-        page = Placeholder();
-        break;
-      case 2:
-        page = Placeholder();
-        break;
-      case 3:
-        page = Placeholder();
-        break;
-      default:
-        throw UnimplementedError('no widget for $selectedIndex');
-    }
-    return Scaffold(
-
-    );
-  }
 }
