+++ miniproyecto/pubspec.yaml
@@ -36,6 +36,7 @@ dependencies:
   cupertino_icons: ^1.0.8
   supabase_flutter: ^2.8.4
   flutter_dotenv: ^5.2.1
+  provider: ^6.1.2
 
 dev_dependencies:
   flutter_test:
